cti_aggregator/backend/requirements.txt
django>=5.1.0
djangorestframework>=3.14.0
django-cors-headers>=4.1.0
django-filter>=23.2
django-celery-beat>=2.5.0
psycopg2-binary>=2.9.6
celery>=5.3.1
redis>=5.0.0
falconpy>=1.2.0
beautifulsoup4>=4.12.2
requests>=2.31.0
pandas>=1.5.0
matplotlib>=3.6.0
plotly>=5.10.0 